DRA-28-add-data-source-to-project

+++ frontend/pages/projects/index.vue
@@ -1,10 +1,28 @@
 <script setup>
-import { onMounted, reactive } from "vue";
+import {useProjectsStore} from '@/stores/projects';
+const projectsStore = useProjectsStore();
 const { $swal } = useNuxtApp();
 const state = reactive({
     project_name: '',
     projects: [],
 })
+watch(
+  projectsStore,
+  (value, oldValue) => {
+    state.projects = projectsStore.getProjects().map((project) => {
+        return {
+            id: project.id,
+            user_id: project.user_platform_id,
+            name: project.name,
+            dataSources: 0,
+            sheets: 0,
+            visualizations: 0,
+            dashboards: 0,
+            stories: 0,
+        }
+    });
+  },
+);
 async function addProject() {
     const inputValue = "";
     const { value: projectName } = await $swal.fire({
@@ -42,6 +60,7 @@ async function addProject() {
                 title: `The project ${projectName} has been created successfully.`,
                 confirmButtonColor: "#3C8DBC",
             });
+            await projectsStore.retrieveProjects();
             getProjects();
         } else {
             $swal.fire({
@@ -51,33 +70,34 @@ async function addProject() {
         }
     }
 }
-async function getProjects() {
-    state.projects = [];
-    const token = getAuthToken();
-    const url = `${baseUrl()}/project/list`;
-    const response = await fetch(url, {
-        method: "GET",
-        headers: {
-            "Content-Type": "application/json",
-            "Authorization": `Bearer ${token}`,
-            "Authorization-Type": "auth",
-        },
-    });
-    const data = await response.json();
-    state.projects = data.map((project) => {
+function getProjects() {
+    state.projects = projectsStore.getProjects().map((project) => {
         return {
             id: project.id,
-            user_id: project.user_platform_id,
             name: project.name,
             dataSources: 0,
             sheets: 0,
             visualizations: 0,
             dashboards: 0,
             stories: 0,
+            user_id: project.user_platform_id,
         }
     });
+    
 }
 async function deleteProject(projectId) {
+    const { value: confirmDelete } = await $swal.fire({
+        title: "Are you sure you want to delete the project?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3C8DBC",
+        cancelButtonColor: "#DD4B39",
+        confirmButtonText: "Yes, delete it!",
+    });
+    if (!confirmDelete) {
+        return;
+    }
     const token = getAuthToken();
     const requestOptions = {
         method: "DELETE",
@@ -94,14 +114,19 @@ async function deleteProject(projectId) {
     } else {
         $swal.fire(`There was an error deleting the project.`);
     }
+    await projectsStore.retrieveProjects();
     getProjects();
 }
+async function setSelectedProject(projectId) {
+    const project = state.projects.find((project) => project.id === projectId);
+    projectsStore.setSelectedProject(project);
+}
 onMounted(async () => {
-    await getProjects();
+    getProjects();
 })
 </script>
 <template>
-    <div class="min-h-100 flex flex-col m-10 border border-primary-blue-100 border-solid p-10 shadow-md">
+    <div class="min-h-100 flex flex-col ml-4 mr-4 md:ml-10 md:mr-10 mt-5 border border-primary-blue-100 border-solid p-10 shadow-md">
         <div class="font-bold text-2xl mb-5">
             Projects
         </div>
@@ -119,30 +144,34 @@ onMounted(async () => {
                     </div>
                 </template>
             </notched-card>
-            <notched-card v-for="project in state.projects" class="justify-self-center mt-10">
-                <template #body="{ onClick }">
-                    <div class="flex flex-col justify-center cursor-pointer">
-                        <div class="text-md font-bold">
-                            {{project.name}}
-                        </div>
-                        <div class="bg-gray-300 p-5">
-                            Screenshot here
-                        </div>
-                        <div class="flex flex-row justify-between mt-1">
-                            <ul class="text-xs">
-                                <li>{{ project.dataSources }} Data Sources</li>
-                                <li>{{ project.sheets }} Sheets</li>
-                                <li>{{ project.visualizations }} Visualizations</li>
-                                <li>{{ project.dashboards }} Dashboard</li>
-                                <li>{{ project.stories }} Story</li>
-                            </ul>
-                            <div>
-                                <font-awesome icon="fas fa-trash" class="text-xl text-red-500 hover:text-red-400" @click="deleteProject(project.id)" />
+            <div v-for="project in state.projects" class="relative">
+                <notched-card class="justify-self-center mt-10">
+                    <template #body="{ onClick }">
+                        <NuxtLink :to="`/projects/${project.id}/data-sources`" class="hover:text-gray-500 cursor-pointer" @click="setSelectedProject(project.id)">
+                            <div class="flex flex-col justify-center">
+                                <div class="text-md font-bold">
+                                    {{project.name}}
+                                </div>
+                                <div class="bg-gray-300 p-5">
+                                    Screenshot here
+                                </div>
+                                <div class="flex flex-row justify-between mt-1">
+                                    <ul class="text-xs">
+                                        <li>{{ project.dataSources }} Data Sources</li>
+                                        <li>{{ project.sheets }} Sheets</li>
+                                        <li>{{ project.visualizations }} Visualizations</li>
+                                        <li>{{ project.dashboards }} Dashboard</li>
+                                        <li>{{ project.stories }} Story</li>
+                                    </ul>
+                                </div>
                             </div>
-                        </div>
-                    </div>
-                </template>
-            </notched-card>
+                        </NuxtLink>
+                    </template>
+                </notched-card>
+                <div class="absolute top-5 -right-2 z-10 bg-gray-200 hover:bg-gray-300 border border-gray-200 border-solid rounded-full w-10 h-10 flex items-center justify-center mb-5 cursor-pointer" @click="deleteProject(project.id)">
+                    <font-awesome icon="fas fa-xmark" class="text-xl text-red-500 hover:text-red-400" />
+                </div>
+            </div>
         </div>
     </div>
 </template>
